requirements.txt
PyPDF2==3.0.1
Pillow==10.2.0
Flask==3.0.0
schedule==1.2.0